convert toSymbolCpp() to D

+++ src/backend/global.h
@@ -322,7 +322,7 @@ void symbol_print(Symbol *s);
 void symbol_term();
 char *symbol_ident(Symbol *s);
 Symbol *symbol_calloc(const char *id);
-Symbol *symbol_calloc(const char *id, size_t len);
+Symbol *symbol_calloc(const char *id, unsigned len);
 Symbol *symbol_name(const char *name, int sclass, type *t);
 Symbol *symbol_name(const char *name, unsigned len, int sclass, type *t);
 Symbol *symbol_generate(int sclass, type *t);
